Add a golang program used to trigger buildkite from gerrit
You need an API key and an account to make changes, but other than
that, it should just work.

It also implements a webhook to detect pass/fail events.  Turns out
buildkite doesn't really like doing anything other than webhooks for
this for what we are doing.

I'm sure I'm not the cleanest on my golang, but it works.

Change-Id: I78707301c26006c73ceee52d2e996bb49e9e7899