Add a bunch of interpolation and cost utilities to NoncausalTimestampFilter
We want to build up an optimization problem which penalizes
 ((ta - tb) - Oa(ta))^2 precisely with doubles.  The optimization
problem will use a gradient decent, so we mostly care about the
derivatives of this cost. This means we need a very accurate way
to compute Oa(ta) and derivatives of the cost.

There are tricks we can play long term by subtracting an int64_t
constant from the optimization problem to drop the high order bits out
of all the calculations to make things more accurate.

Change-Id: I0d6cdc9b3c3c2e98086ace30012e5903a298328d