Add a TimestampProblem class to solve the timestamp estimation problem
The timestamp estimation problem is actually really close to a
MPC.  It can be phrased to have a nice piecewise quadratic cost,
linear constraints, and solves well with similar solvers.  The cost
function chosen is to penalize the squared difference between the
measured and filtered offset, and the difference in the corresponding
times.

The solver chosen in NLOPT, NLOPT_LD_SLSQP, is designed for exactly this
problem type, and uses the gradient to produce a really accurate answer.
Multiple solutions attempted are accurate to full double precision with
gmp.

The next step is to replace the current timestamp estimation algorithm
with this new algorithm.

Change-Id: I7939a97cb845b17ea1fafdf6a09d380cb2cc6d8d