Add a MessageHeader field for monotonic_timestamp_time and extract it
We track the time from send -> republish already by logging delivery
timestamps.  When that time is sent back to the logger node, we actually
measure the time from republish -> logger node too, which would be
useful for timeline reconstruction.  Add a field to log that to and
extract any timestamps we find there.

Next step is to log it and use it again.

Change-Id: I36286903adfa418e830be68382b157794a16ff7a