Track file ending more carefully
The old method was failing on long compressed logs with unclean
shutdown.  We were getting large differences in end times between parts
and CHECKing.  Instead, track when we stop seeing data on a channel, and
when that happens, eat the rest of the log file, confirming that the data
on that channel doesn't come back.  This makes it so we don't try to
determine if we got to the end or not, and instead just confirm that we
don't have gaps in the data.  That is a more direct measurement of what
we actually care about.

Change-Id: I20debdf5623788f66f6c1f01dcb066c8a7ba7ef6