Support inserting times out of order
When we start wanting to use the timestamp on timestamps for the filter,
we will end up with timestamp observations which are out of order.  Find
the right spot in the deque, insert, and then remove points until valid.

Add a *ton* of tests for this.  And while we are here, test that
freezing works too.

Change-Id: If421f0e54628fbd2bd39f25ed8dd58214fc1dac4