Use the timestamp delivery time to reconstruct time
Don't waste the data!

This exposes an issue in sim (and real life).  It takes time between
when a message was sent and when it was forwarded.  Timestamp messages
don't have this delay, so they end up dominating the estimate.  This
causes time estimation to not be absolutely perfect when replaying
simulation, but it is close enough nobody would notice.  Steady state, I
think this is actually benificial behavior.

Change-Id: I85a1e69d47ea09cc61abf82789bcc3b55981f1db