Compress the remote timestamp matching queue
The queue_index_map_ was saving a timestamp for every message ever sent.
This is used to match the queue index that a message was replayed with
to the queue index that is put back in the RemoteMessage to be
re-logged.  This ends up being pretty expensive in terms of memory and
wasn't being cleaned out.

Change-Id: I1d675dec2d782e457304b6f443ebd6c0eac77789