Add drivetrain position plot to drivetrain plotter
It is useful to look at the position plot next to the voltage one
when we are debugging the jerkiness of the drivetrain.
Also, Austin found that we have a bug in drivetrain code by looking
at this plot, so if this commit gets merged then we can all use this
plot to find that bug.

Change-Id: I4c3d58c56154f29849ed761a1d85c123cae0dc2f