Explode when time is unsortable with a better error
So far, this has been shown to be because when you solve for different
nodes, you get slightly different answers.  And those answers can be
un-orderable slightly.

Make it very clear if this is the contributing effect by re-solving for
each node with the bad solution, and show how the different nodes
produce different solutions.  Then, explode.  This is more actionable.

Change-Id: I2e72b49987a835f15197a4dd8a7176a94cde0bff