Make boot_uuid change in event loops when it changes
We were seeing any event loops created before the UUID changes staying
the same.  This defeats the purpose.  This is especially true for the
simulated network bridge which uses an event loop to get the boot UUID,
and wants to observe when it changes.

The boot UUID shouldn't change in any other conditions, so this won't
have an effect on anything else.

Change-Id: Idfa8f36e8d0e4ba964d09bf2170cbbffb17b8276