Refactor starter_cmd a bit
This patch switches the order of arguments for starter_cmd. In
particular, the subcommand now comes first and then you specify the
name of the application.

For example:

  $ starter_cmd start <app>
  $ starter_cmd stop <app>
  $ starter_cmd status <app>

This patch also enhances the "status" command to work without any
arguments. In this case, it will print out a high-level status for all
applications.

The error reporting on incorrect arguments to starter_cmd should also
be quite a bit improved with this change.

Change-Id: I4c80e6662b6f3d3e630a4be26d579e1166a89643