Add voltage error plot and improve drivetrain.py comments
When tuning the 2020/2021 robot, we found that the voltage error was
significantly miss-tuned.  It was resulting in significant oscillations.
To tune it, we need a plot showing the step response.  Make one.

While we were debugging, we found the wording to be confusing for some
of the fields in the parameters.  Update them.

Change-Id: Ibd2f36d1a97305b4703c186a62a90f4b4c289459