Handle a node not having a start time
We don't populate the start time of a log file when we don't know what
time it is on the remote node.  This was leading to us queueing data
relative to the start of time, not the start of the data.

To start with, queue relative to the max of the start time or the
timestamp of the first piece of data.  This will make sure we have
enough data to support sorting.

A follow-up change needs to handle queueing until the start better.

Change-Id: If719a7221d7a3b15992a77d48b398d1344c60c17