Support queueing data in TimestampMapper and viewing timestamps
When solving for the starting time of a log, we don't always have enough
data queued on all the nodes.  We can either do gymnastics to queue
enough data, or we can teach MultiNodeNoncausalOffsetEstimator how to
enqueue more data on each node when asked to evaluate times after the
pre-loaded range.

Teaching MultiNodeNoncausalOffsetEstimator how to automatically enqueue
more data will be more robust in the long term, so let's go that route.

This commit gives us enough hooks to remove the std::deque from inside
LogReader::State and directly use TimestampMapper.  That in turn should
let us automatically enqueue more data when needed, and add it to the
filters automatically.

Change-Id: Ife1248d0b515d9ea9088ace84fa020a351289d0c