Push managing the message queue into TimestampMapper
We want to be able to have multinode_timestamp_filter trigger more data
to be queued.  This will let us manage startup much better when Logger
tries to compute what is happening at the start time (and we aren't
guaranteed to have queued that far ahead).  We can then have
multinode_timestamp_filter grow the queue as timestamps are being
pulled.

Move feeding timetamps into the noncausal filter from LogReader into the
TimestampMapper callback.  This is now registered from inside the
multinode timestamp filter directly.  This also removes the need for
the queue inside LogReader to hold the buffered times.

Change-Id: I054518d9be41d2a6560c45bc3a68a23c8e31347a