Queue filters until there are at least 2 points each
We were seeing issues where each channel had 1 message at start, then a
*long* time gap, then the body of the messages.  We would queue the 1
message up, do time estimation, freeze past that message, then find the
body of data.  That was resulting in us not being able to read
reasonably phrased logs.

The fix is to make sure to queue 2 points for each active filter (plus a
bit after so we don't immediately try to modify it).

Change-Id: I41455bf3c2426233a72f40035430f1bc0c420fdc