Handle missing data at startup better
We had a case where the data was sent *before* the start of the log
file, and the timestamp was sent *after* the start of the log file.  So,
we couldn't match the data because it wasn't logged, but the data wasn't
supposed to be logged whereas the timestamp was supposed to be logged.

The CHECK that fires is trying to catch when data goes away in the
middle of a log and notify the user that we have missing data.  That is
not the case here.

Be more aggressive here and declare any message with missing data where
the data was before the start of the log to be fine and ignore it.
We'll see if there are more cases.

Change-Id: Ia233b0b88699fc3febf2fac1fbd7ee10b03631cd