Handle one node going away better
When one node is never available, we end up setting it's time to the
distributed time for lack of a better time to set it to.  This means
that it is valid, but the canidate solutions end up not being valid.
When we compare, we end up finding times that are before the current
time and freaking out, when we really should have used them earlier.

Ignore any times which are min_time in the solution.  This should only
happen in well defined cases, so we don't need to over-think it yet.

Change-Id: I485891880266825625ad14d3e6895701e5f03a48