Handle finishing a log file more gracefully
Once one file ended, we were spinning until the end and popping
continually.  This was resulting in us removing all the intermediate
data, and then adding it back in with the reverse timestamps.

There are 2 parts to the fix.  1 is detection and the other is fixing
the issue.

1) Instead of tracking a frozen bool, track the frozen time.  This more
   accurately captures what is going on, and handles things being popped
   off the front.  Also, make sure Validate won't certify an answer
   which needs data we popped.

2) When we are spinning to the end of the log, don't forget the
   timestamps yet.  We need those for solving the problem still.
   The current strategy likely keeps too much data at the end, but that
   sure beats keeping too little.

Future improvements should consider managing all of this inside
multinode timestamp filter so we aren't dependent on the logger getting
it right.

Change-Id: I323eff3e19c3b67258835b50a339939360de0444