Add support for both split and combined timestamp channels
In order to handle simulation sending messages as fast as possible, and
messages sent infrequently with timestamps delivered, we want to split
each channel's RemoteMessage timestamps into a separate channel.

Start by creating a class to sort out which one is which reliably, and
start by using and testing it in message_bridge.

Change-Id: I5d279836146a01f5d403f597e59d330440fe702c