Add galactic search challenge ball detection
In the galactic search challenge for the first skills challenge,
the robot has to determine which path is present based on the
positions of the balls in the field, and run that respective spline.
This code has a UI for selecting the regions to look for balls in,
and uses those regions to determine the path. It sends the path
using aos_send to the /camera channel.

Change-Id: Ifd7aacc75eb830b762e14487bad0720dcc2a3944