Use Newton's method to solve our timestamp problem symetrically
We can write out the problem and constraints in such a way that it is
the same problem for all the nodes.  And, this looks faster than nlopt.

I'm too chicken to delete the old nlopt solver and all the
infrastructure around it.  I'll do that if this continues to pan out in
the real world.

Change-Id: I29f9dfb987fb666839424355f75a4dc1f47b57f4