Use overloading instead of default values
As we are getting ready to add the send boot UUID everywhere, that will
require a UUID as an argument to all the Send methods with a default
that is event_loop_->boot_uuid().  That doesn't work.  Instead, use
overloading to achieve the same purpose.

There are few call sites for all of these, so it isn't hard to change
them all.

Change-Id: Id1e17bbef2e63c72fa6ad2b038c69d452610e86e