Break Logger recursion
There are some cases where folks are using the on_logged_period to
rotate and stop the logger.  When they stop the logger, that triggers
the on_logged_period callback to be called again, potentially ending in
an infintate recursion.

Fix that by blocking the callback when flushing at shutdown.

Change-Id: I442b4f5a9d29d996db3e450bd4f652e33ac7feda