Improve galactic search path detection
This will make it look for multiple balls in each layout, and will choose a path
if it finds all of the balls in that layout. Additionally,
the selection of ball regions will be done in a separate program from
the detection loop so that the regions can be reused.

Change-Id: I163af6996a74b37b6f2ef90b188fcd9371add14b