Update galactic search path detection
This has all of the code that we added today while doing the
skills competition, such as eroding and dilating the mask
to amplify the balls and remove noise,
fixing a bug where camera reader was stopped two times,
and updating the json with the rects we selected.
It also adds a feature where you can see the rects you previously
drew on top of the current image.

Change-Id: I1a57d074b4ce05aa39b1fcb9dce4c9c05be057c0