Automatically add split timestamp channels in RemapLoggedChannel
With the new split timestamp channels, we need a corresponding timestamp
channel to be available for timestamps to be sent on for all channels.
When a channel gets remapped, the timestamp channel it's timestamps get
published to changes name.  This makes it so replaying that log file
fails because it can't find the corresponding timestamp channel.

We can require the user to create that channel themselves, but that is
pretty error prone.  Make LogReader do it as part of RemapLoggedChannel
automatically.

While we are here, add a test for the expected behavior.

Change-Id: I1f920b5031c2cedc21f73775797acc6795a7882d