Switch web_proxy over to rawrtc
In theory, nothing else should change...  This will make it possible to
run the web proxy and web plotter in applications which use absl among
other things.

Future work should involve reducing a copy (we copy an extra time when
buffering), and making web_proxy not spin at 100% CPU when it is done by
reshuffling how the event loop works a bit better.

Change-Id: Iab6b6b278c5e830bb71e107824b364dfe7cad46d