Replace EXPECT with ASSERT to avoid container-overflow asan failure
The one in line 1084 is the one that was part of the nightly
failure, the others I added after scanning the file.

1084:  If the size is less than 7 then we get an asan failure
when trying to access elements that don't exist.  The root cause
is not an asan failure though, but a non-deterministic test.
Finding the root cause is beyond me, this change does not address
the root cause.

Change-Id: Ia9fadfaa03636a9c0d49247f7324301aeebc9aac
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>