Populate boot_count for existing logs
This sets us up to figure out which boot comes when so we can sort all
the parts per boot, and then order the boots correctly.

Given that we might get multiple logs from multiple boots from multiple
nodes, we need to be a bit careful how we construct the problem.  The
only way we know how to order a boot from only the data in the header is
to look at remote data from a node that didn't reboot and look at how
the parts_index increases.  Those observations can then be extracted as
constraints and used to generate and sort a DAG, and to check the
connectivity of the DAG to make sure there is 1 valid ordering.

Change-Id: Ic6fadc8538f66c60f3142fcc22d1a91d7822cd1f
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>