Enforce that NodeMerger only operates on one boot
We want to introduce the idea that time can go backwards out a layer
since there is no way to overlap anything between boots when sorting.
CHECK in the constructor that we didn't muck that up.

While we are here, get rid of an un-necesary std::string copy that was
confusing.

Change-Id: Ic26900d5272986ac0c866c58b6b48fb4761e90c1
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>