Add logger boot_count to LogFile when sorting
We need to be able to associate all monotonic times with a boot uuid.
This also includes the timestamps that the timestamps were logged when
they were delivered.  That means we need to know the boot_count for the
logger_node.  Logfile sorting is the best place to do this.

Change-Id: I5f732ca1e9ed001d03d032e0f38241f9d891b044
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>