Add a BootTimestamp class and move some things over
This makes it so we can compare timestamps again across boots.  Move
most things over to using it instead so we can more naturally track what
happens in what order.

Note: this is not complete and has no tests.  Anywhere where I'm not
comfortable that the code will handle it today should now CHECK that the
count is 0 so we explode.  Tests and tolerance of reboots will come in
future patches, starting with TimestampMapper and then with the
time recover code and then LogReader.

Change-Id: I0fdbbe1860de4fe5e48d1a6cd516e7118e5942dc
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>