Work around duplicate headers at the start of logs
Somehow, we have managed to log duplicate headers at the beginning of
some of our logs.
log_11_data/11/aos/aos.message_bridge.Timestamp.part0.bfbs.xz and
friends seem to be where we are finding these issues.  Hide this
behavior behind a flag.

We urgently need a follow-up commit which fixes the log writer so it
doesn't write duplicate headers.  Unfortuantely, we have valuable data
logged with duplicate headers in it so we need this workaround.

Change-Id: I905023fc647547fda5004b303f1a6988ec311a93
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>