Replay logs from both the source and destination nodes
We had logic for falling back to sending remote timestamps immediately
when we found a logged timestamp without the timestamp time.  This was
to handle logs from before we started logging the timestamp time.

This is no longer a case worth worrying about.  Instead, when we find a
message without a timestamp time, that means that the timestamp didn't
make it across the network successfully and the RemoteMessage shouldn't
be recreated and sent.

Change-Id: I6bd44c97b1cabc5fe27c5f8cab6f4929714a90fe
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>