Add node_boots parsing from log file headers
Jim suggested that we could store the boot UUID in the header for all
the nodes rather than per message.  That seems reasonable and lets us do
other things with it in the future.  Use it when available to determine
which boot happens when.

When we only have 1 boot for a node, fall back to setting the boot count
to 0 for sorting.  This lets us get started with the new boot counter.

A follow-up commit will start logging with this.

Change-Id: Ia9e0bd2860de82f70f0daa02782d06f1fd4f4218
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>