Don't queue up the remote when it doesn't exist.
When constructed normally (in log_reader), we call AddPeer.  Based on
how everything is constructed, this sets 'peer' for any pair of nodes
which communicate.  It is impossible to have a set of nodes which only
communicate in one way in AOS today.

In timestamp_extractor, we don't call AddPeer since we don't care to
match and only want local timestamps in order.  Because of this, we
accumulate up all the timestamps and data into the NodeData
datastructure until the program shuts down, exploding our memory usage
for no reason.

Track when we have a peer who cares and don't bother to queue the
messages when nobody cares.  This saves a massive amount of data.

Change-Id: I4b16f31845ce2c14b90994b4ce7e02887dbc60da
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>