Set context to defined value OnRun
The context for the eventloop is currently not set to
anything when entering an OnRun callback. The context
are now set to some known value for simulated and real
time when entering OnRun callback.

While we are here, I noticed that the *remote* uuid was set to the local
UUID in timers.  That makes no sense, so stop doing that.

Change-Id: I2c227a7e9742b092c9079549aad341154460ff11
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>