Make LogNamer copy a template header rather than modify an input
This sets us up much better for making it so LogNamer can decided when
to rotate to handle boot UUIDs changes.  We also have a bug hiding in
log_writer where we get duplicate headers written.  This will make that
a lot harder to do by accident once we've got all the tracking
simplified inside LogNamer.

Change-Id: Ib4a2eb5085e12f74c2c61c58d1f511fc2aeb85fa
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>