Move reboot and header responsibility into DataWriter
Now that we have a class wrapped around the raw flatbuffer writing to
disk, we can start to make it responsible for tracking and writing the
header.  It can also track reboots.  Each message that is to be written
now needs to be passed in with the corresponding boot UUID.

This removes a ton of code in LogWriter since it is simpler to do it in
DataWriter.  We don't need to detect state and then trigger the right
behavior nearly as much because the bottom layers track it very easily
now and just work.

Side effects here of delaying writing a bunch of stuff is that we
shouldn't write headers with no body because we delay writing the header
until we know the body.  This would potentially be a performance problem
with inline configuration, but with the separate configuration files,
this isn't a problem.

Change-Id: Iec91284fcceeb73e3b6c181aebd2fb7edc7bc1ac
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>