Don't create empty parts when a node doesn't boot
Since we delay writing headers until we have data, we create empty files
for log files with no data.  This means that during normal operation, we
create empty files.  This makes it hard to tell if an empty file is
corrupted, or was empty by "design".

Opening a file is required to be cheap enough that we can instead delay
this operation until we get the first message for a log file.

Another option would be to either delete the file on destruction if
empty, or to write the header unconditionally.  I think not writing the
file keeps things simpler for now.

Change-Id: I58156372fc6b30f143690d86a8b7a18822477ac7
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>