Make timestamp_to_csv work with multiple boots
The old code used a starting monotonic time to sync all the data on the
same plot.  This doesn't work when reboots are at play.  Rather than
fight that, use the distributed clock for everything.

Rather than push the concept of the distributed clock into the timestamp
filter, pull the concept of logging CSV files completely into the
multinode filter code.  This actually simplifies things a fair amount.

The next step is to take this beautiful set of boot/time correspondances
and plumb it into the simulation code.

Change-Id: Id9336d683a13dd214eeec3af883d0f38a4cfa711
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>