Change the start time to prefer local logs
We have some logs where the remotely logged channels start before the
local log.  For those, the RT time is min_time.

Thinking back about it, that remote start isn't all that meaningful for
a node if we have a local log.  It typically has a very minimal set of
data that isn't worth parsing.  So we should prioritize local log file
starts when determining the start time.

Change-Id: I1f9795217212d6e9c5d6d74e0abd21973b46db8e
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>