Fix header corruption in log files.
Flatbuffer vectors can change when you touch the fbb.  That is creating
corrupted headers on certain log files.  Stop doing that.

Since we already have logs with the bad headers, make new fields with
the same name but different IDs.  The bad logs will get characterized as
old logs, which is fine.

Change-Id: If08e91a1e30826ff64ff1ecf7223390c50043efa
Signed-off-by: Austin Schuh <austin.linux@gmail.com>