Correctly handle turret-focused image corrections in localizer
This makes it so that, by default, we will prioritize making localizer
corrections with the intent of keeping the turret accurate (note that
this will prioritize whatever image target we are using, even if it
isn't the actual shooting target). Currently, it allows a small portion
of correction to creep through to correct for gyro drift, but is on the
order of <1 deg / minute currently. Given the quality of our gyro, this
should be fine.

Change-Id: I0be93ad434f53a0fc34d2fbcd9889368cc592cb0
Signed-off-by: James Kuszmaul <jabukuszmaul@gmail.com>