Add simple ping pong apps for testing SCTP
We are seeing that when we send a large hunk of data (over 13 ish MB),
we hit SIGXCPU on message_bridge.  Rather than debug with the full AOS
stack in play, I wrote a simple demo app which sends a configurable
message at a configurable period.  It reproduces the issue quite nicely,
for better or worse.

Change-Id: I80681f149efcb21e856e7f65cb0caa4d890c0107
Signed-off-by: James Kuszmaul <james.kuszmaul@bluerivertech.com>