Enforce the actual required rmem and wmem limits we care about
When receiving large messages, the kernel actually starts performing
very poorly with a large rmem.  The only reason to configure a large
rmem is so we avoid fragmentation, which isn't actually possible.  Our
stack supports handling fragmentation already, so there is little harm.
There are ways to make it too small, but you can always configure a
network stack to have really bad performance and AOS shouldn't be in the
buisness of enforcing settings that aren't correctness issues.

wmem needs to be large enough to send the max message size.  So let's
enforce that to avoid the send failure that otherwise occurs.

Change-Id: I49d53285057167e36f1c5b78a4582e30e1d8eb97
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>