Don't crash on channels with timestamps but no data
James tried logging timestamps but not the actual data.  The logger
merrily logged it, but we were unable to read the log.  The code looking
to see if there is supposed to be matching data freaks out because it
doesn't know what the source node is.  We can fill that all in a bit
better with no real extra cost, so let's just do that.

We probably don't want to support doing this at all.  AOS shouldn't let
you create configs like that.

Change-Id: Ia8978a12a5251ce72c271e7bc8c2dcea4c185164
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>