Disallow logging timestamps without data
Nobody has come up with a good use case or a way to read the data
afterwards.  Until we have a use case, it isn't worth spending the
engineering figuring out how to expose this all.

Instead, when validating configs, enforce that we can't log timestamps
without data.  Logs with timestamps and no data are possible to read,
but need --skip_missing_forwarding_entries and are not recommended.

Change-Id: I04534a09cba1bb7f87c0e4e966a3e20ae2007568
Signed-off-by: Austin Schuh <austin.schuh@bluerivertech.com>